feat: refactor URL fetching and save functionality; add flamegraph generation
